tb: coding: move casing primitives to a common module
To be used both in parser and spec.
