+++ crates/tighterror-build/src/common.rs
@@ -1,0 +1,1 @@
+pub mod casing;
