+++ crates/tighterror-build/src/parser/helpers.rs
@@ -1,5 +1,6 @@
 use crate::{
     coder::idents,
+    common::casing,
     errors::{kinds::parser::*, TbError},
     parser::kws,
 };
@@ -19,17 +20,6 @@ pub fn check_ident_chars(ident: &str, desc: &str) -> Result<(), TbError> {
     }
 }
 
-fn convert_case(s: &str, from_case: Case, to_case: Case) -> String {
-    let converter = convert_case::Converter::new()
-        .from_case(from_case)
-        .to_case(to_case);
-    converter.convert(s)
-}
-
-fn is_case(s: &str, case: Case) -> bool {
-    convert_case(s, case, case) == s
-}
-
 fn check_ident(ident: &str, desc: &str, case: Case) -> Result<(), TbError> {
     if ident.is_empty() {
         log::error!("`{desc}` cannot be an empty string");
@@ -38,10 +28,10 @@ fn check_ident(ident: &str, desc: &str, case: Case) -> Result<(), TbError> {
 
     check_ident_chars(ident, desc)?;
 
-    if !is_case(ident, case) {
+    if !casing::is_case(ident, case) {
         log::error!(
             "`{desc}` must be specified in {case:?} case: {ident} -> {}",
-            convert_case(ident, case, case)
+            casing::convert_case(ident, case, case)
         );
         return BAD_IDENTIFIER_CASE.into();
     }
@@ -95,7 +85,7 @@ pub fn check_module_name(name: &str) -> Result<(), TbError> {
     if name.is_empty() {
         log::error!("module name cannot be an empty string");
         BAD_NAME.into()
-    } else if !is_case(name, Case::Snake) {
+    } else if !casing::is_case(name, Case::Snake) {
         log::error!("module name must be specified in lower_snake_case: {name}");
         BAD_NAME.into()
     } else {
@@ -162,25 +152,4 @@ where
     I: IntoIterator<Item = &'a str>,
 {
     check_name_uniqueness("module", iter)
-}
-
-#[cfg(test)]
-mod testing {
-    use super::*;
-    use convert_case::Case::*;
-
-    #[test]
-    fn test_convert_case() {
-        let cases = &[
-            // str_to_convert, from_case, to_case, expected_result
-            ("MY_ERR1", UpperSnake, UpperSnake, "MY_ERR1"),
-            ("MyErr1", UpperSnake, UpperSnake, "MYERR1"),
-            ("MyErr1", UpperCamel, UpperSnake, "MY_ERR_1"),
-            ("MY_ERR1", UpperSnake, UpperCamel, "MyErr1"),
-        ];
-
-        for c in cases {
-            assert_eq!(convert_case(c.0, c.1, c.2), c.3);
-        }
-    }
 }
